Remove PostageClientTest
The test doesn't check anything useful - essentially just that the
PostageClient has the data set that it's meant to (which is either
hard-coded or part of config).

It has been removed as it breaks too easily when there are any
changes to the API URLs and doesn't really help in determining that
the library is working as expected.