Add support for string postcodes
Previously, most postcodes had to be provided as integers, which doesn't
work well for postcodes under 1000, which need to be sent to the API
with the leading 0.

In addition, the requirement for string or integer was inconsistent,
with some calls needing a string only.

For consistency, all API calls which take postcodes as an argument will
now allow integers and strings.

NOTE: Support for integers is deprecated as of this release, and may be
removed entirely in a future major release version (2.x).